+++ packages/ub-pub/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - `AsyncConnection.runTrans` method will add a method names into post request URL (the same as `AsyncConnection.query`),
  so request become `POST /ubql?rq=entity1.method1*entity2.method2`.
+ - AsyncConnection will call `@unitybase/cs-shared.formatByPattern.setDefaultLang` during user session creation.
+ This enables string sorting according to locale and allow calling of `formatByPattern.formatNumber` & `formatByPattern.formatDate`
+ without language parameter (user language will be used).  
  
 ### Changed